requirements.txt
streamlit>=1.22.0
pandas>=1.5.0
requests>=2.28.0
python-dotenv>=0.21.0
numpy>=1.23.0